requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==9.0.2
pytest-django==4.12.0
factory-boy==3.3.3
Faker==40.11.1

# Code quality (optional but recommended)
# black==24.10.0
# flake8==7.0.0
# isort==5.13.2
# pytest-cov==4.1.0
